tests/wpt/tests/css/css-contain/content-visibility/content-visibility-input-image.html

<!DOCTYPE html>
<link rel="author" href="mailto:jarhar@chromium.org">
<link rel="help" href="http://crbug.com/1247417">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div style="content-visibility:hidden" id=container1></div>
<div style="content-visibility:hidden" id=container2></div>

<script>
promise_test(async () => {
  const image1 = document.createElement('input');
  image1.type = 'image';
  const image1Load = new Promise(resolve => {
    image1.addEventListener('load', resolve);
  });
  image1.src = 'resources/dice.png';
  container1.appendChild(image1);
  await image1Load;
  assert_not_equals(image1.width, 0, 'width');

  const image2 = document.createElement('input');
  image2.type = 'image';
  const image2Load = new Promise(resolve => {
    image2.addEventListener('load', resolve);
  });
  image2.src = 'resources/dice.png';
  container2.appendChild(image2);
  await image2Load;
  assert_not_equals(image2.height, 0, 'height');

}, `<input type=image> should return nonzero values for width and height in a c-v:hidden subtree.`);
</script>
